Update main.cpp
When getting root data, the array is checked to see if root data is already inside the array of items to be removed from tree. If it is not, the root data is inserted into the array. Root items are not simply added. If they were, there's a chance that 37 values will not be removed, because of duplicate values.